+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     `prevent_forking_outside_group`, and `file_template_project_id`.
   * Support for `invite_source`, `tasks_to_be_done`, and `tasks_project_id`
     when adding a member to a group or project.
+  * Support `skip_subresources` when deleting group memberships.
 
 ## Deprecations